CustomerMs updates

+++ customerms/src/main/java/com/banking/core/customerms/service/CustomerService.java
@@ -1,20 +1,21 @@
 package com.banking.core.customerms.service;
 
-
 import com.banking.core.customerms.model.CustomerRequest;
 import com.banking.core.customerms.model.CustomerResponse;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.math.BigInteger;
+
 public interface CustomerService {
 
     Flux<CustomerResponse> getAllCustomers();
 
-    Mono<CustomerResponse> getCustomerById(Long id);
+    Mono<CustomerResponse> getCustomerById(BigInteger id);
 
     Mono<String> createCustomer(CustomerRequest request);
 
-    Mono<String> updateCustomer(Long id, CustomerRequest request);
+    Mono<String> updateCustomer(BigInteger id, CustomerRequest request);
 
-    Mono<Void> deleteCustomer(Long id);
+    Mono<Void> deleteCustomer(BigInteger id);
 }
